+++ Configuration/Extbase/Persistence/Classes.php
@@ -6,6 +6,8 @@ return [
     OliverKlee\FeUserExtraFields\Domain\Model\FrontendUser::class => [
         'tableName' => 'fe_users',
         'properties' => [
+            'creationDate' => ['fieldName' => 'crdate'],
+            'modificationDate' => ['fieldName' => 'tstamp'],
             'userGroup' => ['fieldName' => 'usergroup'],
             'lastLogin' => ['fieldName' => 'lastlogin'],
         ],
@@ -13,11 +15,17 @@ return [
     OliverKlee\FeUserExtraFields\Domain\Model\FrontendUserWithCountry::class => [
         'tableName' => 'fe_users',
         'properties' => [
+            'creationDate' => ['fieldName' => 'crdate'],
+            'modificationDate' => ['fieldName' => 'tstamp'],
             'userGroup' => ['fieldName' => 'usergroup'],
             'lastLogin' => ['fieldName' => 'lastlogin'],
         ],
     ],
     OliverKlee\FeUserExtraFields\Domain\Model\FrontendUserGroup::class => [
         'tableName' => 'fe_groups',
+        'properties' => [
+            'creationDate' => ['fieldName' => 'crdate'],
+            'modificationDate' => ['fieldName' => 'tstamp'],
+        ],
     ],
 ];