+++ Tests/Functional/Domain/Repository/FrontendUserRepositoryTest.php
@@ -68,6 +68,8 @@ final class FrontendUserRepositoryTest extends FunctionalTestCase
         $model = $this->subject->findByUid(1);
 
         self::assertInstanceOf(FrontendUser::class, $model);
+        self::assertEquals(new \DateTime('2019-01-01 00:00:00'), $model->getCreationDate());
+        self::assertEquals(new \DateTime('2023-01-01 00:00:00'), $model->getModificationDate());
         self::assertSame('max', $model->getUsername());
         self::assertSame('luif3ui4t12', $model->getPassword());
         self::assertSame('Max M. Minimau', $model->getName());