+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ## x.y.z
 
 ### Added
+- Add `.creationDate` and `.modificationDate` properties (#383)
 - Add support for PHP 8.3 (#382)
 
 ### Changed